requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
cycler==0.10.0
imutils==0.5.3
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
six==1.15.0
toml==0.10.2
wrapt==1.12.1
